Test .yml view (dasel) - Span incorrect
It looks like the behaviour of Vale with dasel view of a Yaml file
gives inconsistent results for Line and Span.

EXPECT:
======
the .Line and .Span records should point to the location in the file that
the .Match text comes from.

ACTUAL:
======
* For .adoc, this expectation is true.

* For rules like Couchbase.VentilatedProse (theory: with `scope: raw`),
we instead get the .Line indexed from the position of the Yaml token selected.

For example, on the first line of the `.description`, it is marked as Line: 1

* For rules like Vale.Terms, we do get the correct .Line, but the .Span is indexed based on *YAML*'s understanding of the parsed content, not the textual source of the .yaml file contents.